+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## [Unreleased][unreleased]
 ### Added
-- More arguments for constructor like Mach no. lambda.
+- Multigridsolver class
+
+## [0.1.1] 2015-03-30
+### Added
+- More arguments for constructor like Mach no., gamma, tau, tolerance level.
 
 ## [0.1.0] 2015-03-30
 ### Added